+++ HISTORY.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.3.8]
+### Fixed
+- Error message from firebase init
+- Images deploy properly on publish
+- Improvements to config read for deploy/publish 
+
+### Added
+- Error resolution text
+
+## [1.3.7]
+### Added
+- ES6 passes deploy/publish linter
 ## [1.3.4]
 ### Changed
 Properly handles bad searches